+++ src/ebird/api/data/admin.py
@@ -13,7 +13,7 @@ from django.views import generic
 from . import models
 from .fields import TranslationCharField, TranslationTextField
 from .loaders import APILoader
-from .models import Observation, Species
+from .models import Observation, Species, Checklist
 
 
 class ObservationInline(admin.TabularInline):
@@ -44,6 +44,17 @@ class ObservationInline(admin.TabularInline):
             .order_by("species__taxon_order")
         )
 
+class ProtocolListFilter(admin.SimpleListFilter):
+    title = _("protocol")
+    parameter_name = "protocol_code"
+
+    def lookups(self, request, model_admin):
+        return Checklist.Protocol.choices
+
+    def queryset(self, request, queryset):
+        if value := self.value():
+            return queryset.filter(protocol_code=value)
+
 
 @admin.register(models.Checklist)
 class ChecklistAdmin(admin.ModelAdmin):
@@ -55,6 +66,7 @@ class ChecklistAdmin(admin.ModelAdmin):
         "location",
         "observer",
     )
+    list_filter = [ProtocolListFilter]
     ordering = ("-started",)
     search_fields = ("identifier", "location__name", "observer__name")
     autocomplete_fields = ("location", "observer")